requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
mysql-connector-python
python-dotenv
pydantic
pytest